inner loop is working
